Species allocation python script
This script will use the output from a SKA anaylsis, and determine which isolates are within the threshold distance, and group them together.  It outputs a file than be printed to a .csv file to be used for further analysis.  

The threshold is not a hard cut-off but allows to isolates to be clustered together when they are within the threshold distance of some of the other samples in the group.